+++ client/components/App.jsx
@@ -1,14 +1,14 @@
 import React, { useState } from 'react'
 import { BrowserRouter as Router, Route, useParams } from 'react-router-dom'
 
-import Album from './Album'
-import Home from './Home'
+// import Album from './Album'
+// import Home from './SearchResult'
 import NavBar from './NavBar'
 import Footer from './Footer'
 import UserDialog from './UserDialog'
-import Routing from './Routing'
-import Search from './Search'
-import Testing from './Testing'
+// import Routing from './Routing'
+import Home from './Home'
+import SearchResult from './SearchResult'
 
 import CssBaseline from '@mui/material/CssBaseline'
 import { createTheme, ThemeProvider } from '@mui/material/styles'
@@ -88,10 +88,10 @@ const App = () => {
           handleRegisterOpen={handleRegisterOpen}
         />}
         <Route exact path="/">
-          <Search />
+          <Home />
         </Route>
-        <Route path="/noodle/:value">
-          <Testing />
+        <Route path="/noodle/:searchTerm">
+          <SearchResult />
         </Route>
         <Footer />
       </Router>
