+++ client/components/Home.jsx
@@ -1,36 +1,38 @@
 import React, { useState, useEffect } from 'react'
-import { getRecipes, getFaveRecipes } from '../apis/recipes'
-
-import { IfAuthenticated, IfNotAuthenticated } from './Authenticated'
-import NoodleCard from './NoodleCard'
 
 import Box from '@mui/material/Box'
 import Container from '@mui/material/Container'
 import Grid from '@mui/material/Grid'
 import Typography from '@mui/material/Typography'
+import Card from '@mui/material/Card'
+import CardActionArea from '@mui/material/CardActionArea'
+import CardContent from '@mui/material/CardContent'
 
-function Home ({ faves }) {
-  const [noodles, setNoodles] = useState(null)
+const noodleTypes = [
+  "Ramen",
+  "Fettuccine",
+  "Soba",
+  "Vermicelli",
+  "Rice Sticks",
+  "Spaghetti",
+  "Hokkien Noodles",
+  "Egg Noodles",
+  "Udon",
+  "Tagliatelle",
+  "Lo Mein",
+  "Glass Noodles",
+  "Bucatini",
+  "Linguine",
+  "Pappardelle"
+  ]
 
-  const randomNoodles = () => {
-    return getRecipes('noodles')
-      .then(data => setNoodles(data))
-  }
-
-  const showFaves = () => {
-    getFaveRecipes()
-      .then(data => setNoodles(data))
-      .catch(err => console.log(err.message))
-  }
-
-  useEffect(() => {
-    faves ? showFaves() : randomNoodles()
-  }, [faves])
+function Home () {
 
   return (
     <Box
       sx={{
-        py: 5
+        py: 5,
+        minHeight: '90vh'
       }}
     >
       <Container maxWidth="sm" >
@@ -46,31 +48,28 @@ function Home ({ faves }) {
           color="text.secondary"
           gutterBottom
         >
-        Check out the recipes below for how to use your noodle.
+        Click on your noodle to see how to use it.
         </Typography>
-        <IfAuthenticated>
-          <Typography
-            variant="h5"
-            color="text.primary"
-            gutterBottom
-          >
-          Click the heart to save to your favourites.
-          </Typography>
-        </IfAuthenticated>
-        <IfNotAuthenticated>
-          <Typography
-            variant="h5"
-            color="text.primary"
-            gutterBottom
-          >
-          Register to save recipes as favourites.
-          </Typography>
-        </IfNotAuthenticated>
       </Container>
-      <Container sx={{ py: 5 }} maxWidth="lg">
+      <Container sx={{ py: 5 }} maxWidth="md">
         <Grid container spacing={2}>
-          {noodles?.map((item) => (
-            <NoodleCard noodle={item} key={item.uri || item.externalId}/>
+          {noodleTypes.map((item, index) => (
+            <Grid item xs={12} sm={6} md={4} lg={4} key={index}>
+              <Card sx={{
+                height: '100%',
+                display: 'flex',
+                flexDirection: 'column',
+                backgroundColor: 'secondary.main'
+              }}>
+                <CardActionArea href={`/noodle/${item}`}>
+                  <CardContent>
+                    <Typography variant="h5" align="center">
+                      {`${item}`}
+                    </Typography>
+                  </CardContent>
+                </CardActionArea>
+              </Card>
+            </Grid>
           ))}
         </Grid>
       </Container>
